+++ examples-pydantic-v1/windmill_pydantic_v1/_api/windmill_query.py
@@ -220,7 +220,6 @@ class WindmillQueryAPI(QueryAPI[T_DomainModelList]):
         return self._query()
 
     def _query_append_nacelle(self, from_: str) -> None:
-        view_id = Nacelle._view_id
         self._builder.append(
             NodeQueryStep(
                 name=self._builder.create_name(from_),
@@ -228,14 +227,13 @@ class WindmillQueryAPI(QueryAPI[T_DomainModelList]):
                     from_=from_,
                     through=self._view_id.as_property_ref("nacelle"),
                     direction="outwards",
-                    filter=dm.filters.HasData(views=[view_id]),
+                    filter=dm.filters.HasData(views=[Nacelle._view_id]),
                 ),
                 result_cls=Nacelle,
             ),
         )
 
     def _query_append_rotor(self, from_: str) -> None:
-        view_id = Rotor._view_id
         self._builder.append(
             NodeQueryStep(
                 name=self._builder.create_name(from_),
@@ -243,7 +241,7 @@ class WindmillQueryAPI(QueryAPI[T_DomainModelList]):
                     from_=from_,
                     through=self._view_id.as_property_ref("rotor"),
                     direction="outwards",
-                    filter=dm.filters.HasData(views=[view_id]),
+                    filter=dm.filters.HasData(views=[Rotor._view_id]),
                 ),
                 result_cls=Rotor,
             ),