+++ docs/CHANGELOG.md
@@ -14,6 +14,19 @@ Changes are grouped as follows
 - `Security` in case of vulnerabilities.
 
 ## TBD
+### Added
+- Support for reverse direct relations. This includes
+  - All read data classes now include reverse direct relations fields.
+  - In the `.list` method a `retrieve_connections` parameter is added that can be set to
+    `full` to retrieve all reverse direct relations, direct relations, and edges with
+    destination node for each item in the returned list.
+
+### Removed
+- In the `.list` method, the `retrieve_edges` parameter is removed. This has been replaced
+  by the `retrieve_connections` parameter. Setting the `retrieve_connections="identifier'`
+  will have the same behavior as `retrieve_edges=True` and `retrieve_connections="skip"`
+  the same as `retrieve_edges=False`.
+
 ### Fixed
 - When querying, in the unpacking, the generated SDK no longer assumes that all
   edges pointing out/in of a node have unique edge type.